Update file_actor for hyrax 3, and switch it to an override
